requirements.txt
vk_api==11.9.9
python-dotenv==1.0.1
requests==2.31.0

google-generativeai
